Fix build of qiconvcodec.cpp
When configuring with cmake on macOS or Ubuntu 20.04, we don't seem to
detect the posix-compliance of iconv correctly. Since it's acceptable
to use a const_cast here anyway, we might just as well always do so to
pass a plain char* to the iconv function; if it expects a const char*,
then that's ok; the other way around is not.

Change-Id: I803c92af8aa8249514c12150a71dac7f93c6a5e1
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>