configure: Merge the three iconv libraries
The distinction between the GNU and POSIX flavors is not necessary
anymore since a recent commit that employs a const_cast to support both
flavors unconditionally.

The SUN iconv library is the same as GNU without the -liconv linker
flag. The iconv versions that need or may not need the -liconv linker
flag are covered by one common "iconv" library.

Fixes: QTBUG-84708
Change-Id: I3db530c9403afd9bc3db43849f7da1b4fcda917c
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>