Remove superfluous platform and feature check
The variable 'codeset' is only used in iconv_open calls. There's no
signature difference between GNU, POSIX, QNX and SunOS variants for
the iconv_open function, therefore the type of 'codeset' must be
'const char *' in all cases.

This change is in preparation of merging the three iconv features in a
subsequent commit.

Change-Id: Ibe609a19028cdc03668aae496c7d798aa71db246
Reviewed-by: Kai Koehne <kai.koehne@qt.io>