Remove the additional Qt5 namespace
The additional Qt5 namespace in Qt5Compat requires projects using
classes from there not only to touch the buildsystem to link to the
module but also the code itself. This hampers migration from Qt5 to Qt6.

Note that adding a plain 'using namespace Qt5;' in some central header is
not sufficient as a work around when using QTextCodec.

Task-number: QTBUG-85504
Change-Id: Id179663b70dd2c5f45e2e09f3dfac53377faa380
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Sona Kurazyan <sona.kurazyan@qt.io>