Move graphical effects to qt5compat
To assist with porting from Qt 5, we move the Qt Graphical
Effects into Qt5Compat. Mainly, this is copying files and
changing the import name to be Qt5Compat.GraphicalEffects.

A few effects depend on dynamic generation and have thus been
removed in the source repository, as this is not compatible with
how we do shaders in Qt 6. These are now listed in documentation.

Adds a manual test which I used to verify the effects, and which
might be useful for quick testing.

[ChangeLog][Qt Graphical Effects] Added version of Qt Graphical
Effects for compatibility with Qt 5 code. This can be accessed
by importing "Qt5Compat.GraphicalEffects"

Change-Id: I6f63e61ebd282fcf40da8e8e62167d13bf83f371
Reviewed-by: Paul Wicking <paul.wicking@qt.io>