Fix QTextCodec::canEncode() for ICU codec
QTextCodec::canEncode() relies on the number of invalid characters
to determine if the encoding is possible or not.

By default the ICU fromUnicode converter method does not provide any
ways to track the amount of failures. However it uses callbacks to
report errors or replace unrecognized characters with substitute string.

This patch introduces a custom callback for fromUnicode conversion.
The callback just increases the invalid characters counter and then
calls the default callback, which does its usual job.

Pick-to: 6.2
Task-number: QTBUG-83081
Change-Id: Ic2c10285777128a25668551db2a0de4a1b9815a8
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>