Update commercial license headers
Updated header.COMM to the files in tqtc-qt5compat.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4932
Change-Id: Ifd2cef68f881febea6638bb3ef4d1c95e0408fbc
Reviewed-by: Sona Kurazyan <sona.kurazyan@qt.io>