QText{En,De}coder: use DefaultConversion
In Qt 5, the flag to write the BOM was a negative: if IgnoreHeader
wasn't set, we would write it. That was the default in Qt 5 because
DefaultConversion was 0. But now it's a positive flag, DefaultConversion
was updated to set it, but we forgot to use it in the constructors of
these two classes.

[ChangeLog][Qt5CoreCompat][QTextEncoder] Fixed a bug that caused
QTextEncoder not to write the Byte Order Mark for UTF codecs when the
constructor without explicit flags was used.

Fixes: QTBUG-122795
Pick-to: 6.5 6.7
Change-Id: If1bf59ecbe014b569ba1fffd17c459af4e0e00c9
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>