Doc: Fix incorrect usage of \generatelist
The \generatelist command takes an optional argument
that affect how the members are listed, for example,
`classesbymodule` which creates an annotated list of
the classes in a named C++ module.

`groupsbymodule` is not recognized or documented. QDoc
falls back to generating a list of all members of a
group in this case. While the end result is correct,
this relies on an undocumented feature that is likely
to change and break.

Replace these instances with an explicit \annotatedlist
command that produces the same output.

Task-number: QTBUG-138901
Change-Id: I4486d19f59c6c1fd085f15a6a296226a6a8a20b0
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 533e758dba7c93d65bdaa8c5553455581123f4da)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>