Inline cp949codetbl_p.h into its only user
There doesn't seem to be a generator that creates this file, so that
can't be the reason to have it as a separate header.

There's also only qeuckrcodec.cpp which includes the header, so
multiple users, too, can't be the reason.

Since there's no reason to have it as a separate header, don't. Inline
the contents into qeuckrcodec.cpp.

As a drive-by, make the table constexpr instead of just const.

This avoids having to mark the file with Qt-Security headers.

Task-number: QTBUG-135745
Pick-to: 6.10 6.8
Change-Id: Ie0487bc62ab2e88727bbfa442ab85549670bbf57
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>