tst_QLinkedList: fix a comment
We're removing an element with value 20. It's also element number
three in C counting (it's the fourth element in natural counting).

I don't know what the comment meant with "2", so remove it, and
indicate in the append() chain which element we're erasing.

Amends the move of the test to qt5compat (but probably the start of
the public history).

Pick-to: 6.10 6.8 6.5
Change-Id: I7362da1f75e5bed272405e89c4139fd9696333e6
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
(cherry picked from commit bde152a7a98404566d37528cd11921abafbdce8e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>